Fix issues with pcv getting-started guide
Fix the issues with the getting-started guide for pluscloud VMware
that were identified during review of the pull request to the official
documentation repository.

These were mainly typos and some formatting issues where quotes or
Buttons were not formatted correctly.

Also an outdated screenshot of the old customer portal was removed
because on the one hand it was not useful anymore and on the other
hand it showed where IP adresses were to be found which is now
included in the vCloud Director Web UI whenever this info is needed.

Signed-off-by: Sascha Scherrer <sascha.scherrer@plusserver.com>