Modify image guidelines
As it turned out, including the images as base64 encoded string inside
the markdown file makes reviews uncomfortable because there is a lot of
what seems like text and there are short text elements in between.

Furthermore, the GitHub WebUI diff view when reviewing PRs refuses to
display such large text files.

Hence it would be better to allow for import of images and to have
them in a consistent manner, we introduce a shortcode to refer to an
image.

Signed-off-by: Sascha Scherrer <sascha.scherrer@plusserver.com>